Move barcode service to fridge and add type info
