+++ backend/src/routes/routes.ts
@@ -7,7 +7,6 @@ import mediaRoutes from './media';
 import usersRoutes from './user';
 import foodItemRoutes from './foodItem';
 import foodTypeRoutes from './foodType';
-import barcodeRoutes from './barcode';
 import recipeRoutes from './recipe';
 import notificationRoutes from './notification';
 import fridgeRoutes from './fridge';
@@ -18,7 +17,6 @@ router.use('/auth', authRoutes);
 router.use('/hobbies', authenticateToken, hobbiesRoutes);
 router.use('/user', authenticateToken, usersRoutes);
 router.use('/media', authenticateToken, mediaRoutes);
-router.use('/barcode', authenticateToken, barcodeRoutes);
 
 router.use('/fridge', authenticateToken, fridgeRoutes);
